server/requirements-ocr.txt
paddleocr
paddlepaddle
numpy
Pillow
